change id type

+++ offset/database/offset.go
@@ -1,7 +1,7 @@
 package database
 
 type Offset struct {
-	ID        uint64
+	ID        uint
 	Topic     string
 	Partition int32
 	Offset    int64
